refactor: extract timeline type enrichment
